# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Integration Google Calendar",
        "version": "0.0.1"
    },
    "paths": {
...
        "api/google/credentials/register-key": {
            "post": {
                "tags": [
                    "Credentials Google"
                ],
                "summary": "Cria credenciais de configuração",
                "description": "A rota realiza a criação de credenciais de configurações da integração",
                "operationId": "createCredential",
                "requestBody": {
                    "description": "User credentials config google",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "google_calendar_id": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "google_client_id": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "google_client_secret": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "google_redirect_uri": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "google_webhook_uri": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "credential": {
                                            "$ref": "#/components/schemas/GoogleCredentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "api/google/credentials/list-keys": {
            "get": {
                "tags": [
                    "Credentials Google"
                ],
                "summary": "Lista Credenciais",
                "description": "A rota realiza a listagem de credenciais associados ao usuário conectado",
                "operationId": "listCredentials",
                "responses": {
                    "204": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ops, no credentials registered"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GoogleCredentials"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "api/google/credentials/update-key": {
            "put": {
                "tags": [
                    "Credentials Google"
                ],
                "summary": "Atualiza Credencial",
                "description": "A rota realiza a atualização da credencial associada ao usuário",
                "operationId": "updateCredential",
                "parameters": [
                    {
                        "name": "credentialId",
                        "in": "path",
                        "description": "Id da credencial",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Please enter credentialId"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ops, an error occurred when performing the update"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Credential successfully updated"
                                        },
                                        "color": {
                                            "$ref": "#/components/schemas/GoogleCredentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "google/credentials/delete-key": {
            "delete": {
                "tags": [
                    "Credentials Google"
                ],
                "summary": "Deleta Credencial",
                "description": "A rota realiza a deleção da credencial associada ao usuário",
                "operationId": "deleteCredential",
                "parameters": [
                    {
                        "name": "credentialId",
                        "in": "path",
                        "description": "Id da credencial",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Please enter credentialId"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The credential reported does not exist"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Credential successfully deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GoogleCredentials": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "google_calendar_id": {
                        "type": "text",
                        "readOnly": "true"
                    },
                    "google_client_id": {
                        "type": "text",
                        "readOnly": "true"
                    },
                    "google_client_secret": {
                        "type": "text",
                        "readOnly": "true"
                    },
                    "google_redirect_uri": {
                        "description": "Redirection route evoked by google on Oauth2 login",
                        "type": "text",
                        "readOnly": "true"
                    },
                    "google_webhook_uri": {
                        "description": "Redirection route evoked by when there is a change in the event",
                        "type": "text",
                        "readOnly": "true"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "GoogleCredentials"
                }
            },
            "User": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "John Due"
                    },
                    "email": {
                        "description": "User unique email address",
                        "type": "string",
                        "format": "email",
                        "readOnly": "true",
                        "example": "user@gmail.com"
                    },
                    "provider_id": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "1051527092716338246212"
                    },
                    "provider_name": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "google"
                    },
                    "google_access_token_json": {
                        "type": "string",
                        "readOnly": "true"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}